requirements.txt
certifi==2017.4.17
chardet==3.0.4
contentful-management==1.3.1
idna==2.5
invoke==0.19.0
py==1.4.34
pytest==3.1.3
python-dateutil==2.6.0
PyYAML==3.12
requests==2.18.1
six==1.10.0
urllib3==1.21.1
